// src/api/types/SearchCorporaParameters.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

import * as Vectara from "../index.js";

/**
 * The parameters to search one or more corpora.
 */
export interface SearchCorporaParameters extends Vectara.SearchParameters {
    /** The corpora that you want to search. */
    corpora?: Vectara.KeyedSearchCorpus[];
}